Disable tar.gz archives
tar.gz archives the files first then compresses the archive.
This prevents the binary diff from working very well resulting
in large diffs for very small changes.

The tradeoff here is that zip loses permissions. So the restart command
needs to be updated to change permissions on the app before restarting